+++ CPU_IDE/inc/fb_code/fb_user/fcZ1_64_OVG.h
@@ -36,7 +36,7 @@ typedef struct {
 	float					t1;  				// Time 1 
 		
     // Output
-    float		 			o1_alarm;			// Out1 Alarm	
+    uint32		 			o1_alarm;			// Out1 Alarm	
 } strFC0703Info;
 
 #endif
