+++ CPU_IDE/inc/fb_code/fb_user/fcZ2_50N51N_TEOCP.h
@@ -39,8 +39,8 @@ typedef struct {
 	float 					t2;  				// Time 2 
 		
     // Output
-    float 					o1_alarm;			// Out1 Alarm
-    float					o2_conBlk;			// Out2 conBlk
+    uint32 					o1_alarm;			// Out1 Alarm
+    uint32					o2_conBlk;			// Out2 conBlk
 } strFC0709Info;
 
 #endif
