+++ CPU_IDE/inc/fb_code/fb_user/fcZ4_60APN.h
@@ -40,8 +40,8 @@ typedef struct {
 	float			t2; 				// Time 2
 		
     // Output
-    float  			o1_alarm;			// Out1 alarm
-    float  			o2_conBlk;			// Out2 conBlk
+    uint32  		o1_alarm;			// Out1 alarm
+    uint32  		o2_conBlk;			// Out2 conBlk
 } strFC0731Info;
 
 #endif
