+++ CPU_IDE/inc/fb_code/fb_user/fcZ4_76D_DOCP.h
@@ -38,8 +38,8 @@ typedef struct {
 	float		 			t1_low;				// Time 1 Low
 	
     // Output
-    float  					o1_emst_up;			// Out1 EMST Up
-    float  					o1_emst_low;		// Out1 EMST Low
+    uint32  				o1_emst_up;			// Out1 EMST Up
+    uint32  				o1_emst_low;		// Out1 EMST Low
 } strFC0728Info;
 
 #endif
